Mega merge de Vision previo a refactorización completa:
Audio
- Actualizado FMOD
- Bug memory leak en BaseSound.h
- SoundFmod: added isLoaded y getDuration

CameraInput
- showFPS (of camera capture)
- VICamera multithreading by parameter

ComputerVision
- added TrackedBlob

ExternLibs
- PortVideo reorganizing to make it work again
- Aded TinyXML but not in use by vision yet

Framework
- finishApp to finish application execution
- keyboard accesible from user application

Graphics
- DynamicLines added
- line function added
- ImageInkModes added
- Object3D: can receive parent scene node in init
- TexturedQuad::setAdditiveMode

GUI
- Keyboard events injected to GUI
- Keyboard::isModifierDown

+++ vision/src/common/MathUtils.h
@@ -244,7 +244,7 @@ struct Average
 	}
 
 	/// @brief ads a value to average
-	void addValue( float value )
+	void addValue( double value )
 	{
 		if ( values.size() < nValues )
 			values.push_back( value );
@@ -258,11 +258,11 @@ struct Average
 	/// @brief returns the ave
 	float getValue()
 	{
-		float sum = std::accumulate( values.begin(), values.end(), 0 );
-		return sum / (float)values.size();
+		double sum = std::accumulate( values.begin(), values.end(), 0.0 );
+		return sum / (double)values.size();
 	}
 
-	std::vector< float >	values;
+	std::vector< double >	values;
 	size_t								nValues;
 	size_t								index;
 };
